sparc, m68k: wire up tests/c/vlong and tests/c/float, fix a real Ieee struct bug
Both architectures pass byte-exact, including a genuine kc/kl and
2c/2l bug found along the way (not a hand-written-stub issue):
include/obj/k.out.h and include/obj/2.out.h declared the Ieee struct's
float-constant halves as `long l; long h;` instead of `int32 l; int32
h;` like every other working arch's own *.out.h. `long` is 8 bytes on
this 64-bit host, but the linker's D_FCONST reader/writer (obj.c/
asm.c) treats the struct as a tightly-packed 8-byte blob -- so any
FCONST double whose low mantissa byte has bit 7 set got sign-extended
garbage into its own high half. Found via tests/c/float: 2.2 printed
as NaN, and even the FREGZERO/HALF/ONE/TWO hardwired constant
registers linux_sparc.s/linux_power.s initialize came out wrong (0.5
loaded as 0.0, cascading into 2.0/3.0 also reading as 0.0), since
loading 0.5 itself goes through the same broken path.

tests/c/vlong needed one more fix: vlrt.c's byte-order-dependent
struct Vlong only had #ifdef mips/#ifdef power branches for
big-endian; added #ifdef sparc and #ifdef m68k (both big-endian) to
the same nesting.

Checked a real Plan9 distribution's libc/{sparc,68020}/ for reusable
code (muldivrt.s, vlop.s, vlrt.c) per suggestion, but kc/kl and 2c/2l
already carry everything these two tests need -- kept as a reference
for future bring-up work. This repo's own naming stays "sparc"/"m68k"
throughout, never Plan9's "68020".

Full writeup in docs/claude_notes/notes_arch_sparc.txt and
notes_arch_m68k.txt.

+++ include/objexec/2.out.h
@@ -516,8 +516,15 @@ enum
 typedef	struct	ieee	Ieee;
 struct	ieee
 {
-	long	l;	/* contains ls-man	0xffffffff */
-	long	h;	/* contains sign	0x80000000
+	/* claude: int32, not long -- see k.out.h's own comment on this
+	 * exact struct (found there first): a `long` field here is 8
+	 * bytes on a 64-bit host, and the byte-copy/byte-shift code in
+	 * this arch's own linker treats &ieee as a tightly-packed 8-byte
+	 * blob, so the wider type corrupts any FCONST double whose low
+	 * mantissa byte has bit 7 set. See q.out.h/v.out.h/i.out.h/
+	 * 6.out.h, which already use int32 here. */
+	int32	l;	/* contains ls-man	0xffffffff */
+	int32	h;	/* contains sign	0x80000000
 				    exp		0x7ff00000
 				    ms-man	0x000fffff */
 };
